vst-plugin/python-backend/requirements.txt
# text2midi VST3 Plugin — Python Backend Dependencies
# ── Server ──────────────────────────────────────────
fastapi>=0.100.0
uvicorn[standard]>=0.25.0

# ── Pipeline (same as root requirements.txt) ────────
langgraph>=0.1.0
langchain>=0.1.0
langchain-groq>=0.1.0
langchain-core>=0.1.0
openai>=1.0.0
mido>=1.3.0
music21>=9.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
numpy>=1.24.0
tenacity>=8.2.0
typing-extensions>=4.7.0
platformdirs>=4.0.0

# ── Build (optional) ────────────────────────────────
# pyinstaller>=6.0.0
